reset styles for a:not([href])
See https://github.com/twbs/bootstrap/pull/19411 for reference

Having different styles for <a> with and without href is the default
browser behavior.

We could also do link styling on a[href], but that would be hard to
overwrite.

The downside of this solution is that it is now hard to overwrite
a:not([href]), but that is a much less common task.